modulize code
