+++ index.css
@@ -4,29 +4,21 @@
   box-sizing: border-box;
 }
 
+::-webkit-scrollbar {
+  width: 1em;
+}
+
+::-webkit-scrollbar-thumb {
+  background-color: darkgrey;
+}
+
 body {
   margin: 0;
   font-family: monospace;
   height: 100vh;
   width: 100vw;
-  display: grid;
-  grid-template-areas:
-    "header"
-    "terminal"
-    "guide"
-    "footer";
-  grid-template-columns: auto;
-  grid-template-rows: 30px 1fr 1fr 30px;
-}
-
-@media screen and (min-width: 600px) {
-  body {
-    grid-template-areas:
-      "header header"
-      "terminal guide"
-      "footer footer";
-    grid-template-rows: 30px 1fr 30px;
-  }
+  display: flex;
+  flex-direction: column;
 }
 
 .body--dark {
@@ -36,13 +28,18 @@ body {
 
 header {
   padding: 8px;
-  grid-area: header;
 }
 
 main {
   padding: 8px;
   overflow-y: auto;
-  grid-area: terminal;
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column-reverse;
+}
+
+.cmd {
+  flex-grow: 1;
 }
 
 .cmd__prompt {
@@ -62,10 +59,6 @@ main {
   opacity: 0;
 }
 
-aside {
-  grid-area: guide;
-}
-
 iframe {
   width: 100%;
   height: 100%;
@@ -74,5 +67,4 @@ iframe {
 footer {
   padding: 8px;
   color: #888888;
-  grid-area: footer;
 }
