+++ index.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>CLI</title>
-
+    <title>Yet Another Command Guide</title>
+    <link rel="shortcut icon" type="image/png" href="./favicon.png" />
     <link rel="stylesheet" href="./index.css" />
   </head>
   <body class="body--dark">
@@ -13,19 +13,16 @@
       🤖 Yet Another Command Guide
     </header>
     <main>
-      <div class="terminal-output">
-        Welcome!
-      </div>
       <div class="cmd">
         <span class="cmd__prompt">></span><span class="cmd__output"></span
         ><span class="cmd__cursor">&nbsp;</span>
         <textarea class="cmd__input" autofocus></textarea>
       </div>
+      <div class="terminal-output">
+        Welcome!
+      </div>
     </main>
-    <aside>
-      <iframe id="guide" src="" frameborder="0"></iframe>
-    </aside>
-    <footer>Commands: <code>ls</code> <code>cat</code></footer>
-    <script src="./index.js"></script>
+    <footer>Commands: <span id="command-list"></span></footer>
+    <script type="module" src="./index.js"></script>
   </body>
 </html>
